python/requirements.txt
pyyaml~=6.0
# yaml~=0.2.5
seaborn~=0.11.2
joblib>=1.2.0
scikit-learn~=1.0.1
numpy>=1.21.2
pandas
matplotlib
gradio